// enum.ts
export enum ItemType {
    Clothing = "Clothing",
    Accessories = "Accessories",
    Electronics = "Electronics",
}

export enum DiscountType {
    FixedAmountDiscount = "FixedAmountDiscount",
    PercentageDiscount = "PercentageDiscount",
    PercentageDiscountByCategory = "PercentageDiscountByCategory",
    DiscountByPoint = "DiscountByPoint",
    SpecialCampaignDiscount = "SpecialCampaignDiscount",
}